+++ src/dpsprep/exceptions.py
@@ -6,5 +6,9 @@ class DpsPrepError(DpsPrepException):
     pass
 
 
+class DpsPrepParseError(DpsPrepError):
+    pass
+
+
 class DpsPrepConfigError(DpsPrepError):
     pass